adding modularized tests
