--- mesh/helloworld.c
@@ -1,6 +1,0 @@
-#include "helloworld.h"
-#include <stdio.h>
-
-void say_hello() {
-    printf("Hello World!\n");
-}
