--- mesh/include/helloworld.h
@@ -1,6 +1,0 @@
-#ifndef HELLOWORLD_H
-#define HELLOWORLD_H
-
-void say_hello();
-
-#endif
