+++ mesh/test/shmem_handler/test_writer.c
@@ -9,7 +9,7 @@
 #include <sys/mman.h>
 #include <sys/stat.h>
 
-#include "mesh.h"
+#include "shmem_handler.h"
 
 void test_shmget_should_return_fd() {
     printf("### test_shmget_should_return_0\n");
@@ -17,16 +17,6 @@ void test_shmget_should_return_fd() {
     assertm(shm_fd>=0, "Reserve memory found an error");
 }
 
-void test_shmem_should_close() {
-    printf("### test_shmem_should_close\n");
-    open_shared_memory(TEST_SHM_NAME);
-
-    int res = close_shared_memory(TEST_SHM_NAME);
-    assertm(res==0, "Closing shared memory found an error");
-
-    int reopening_errcode = shm_open(TEST_SHM_NAME, O_RDWR, 0666);
-    assertm(reopening_errcode<0, "Shared memory could still be openned after supposedly closing it");
-}
 
 void test_should_give_a_size_to_shmem() {
     printf("### test_should_give_a_size_to_shmem\n");
@@ -40,13 +30,30 @@ void test_should_give_a_size_to_shmem() {
 void test_should_retrieve_ptr_from_shmem() {
     printf("### test_should_retrieve_ptr_from_shmem\n");
     int shm_fd = open_shared_memory(TEST_SHM_NAME);
-    give_size_to_shmem(shm_fd, sizeof(int));
+    size_t size = sizeof(int);
+    give_size_to_shmem(shm_fd, size);
 
-    void *shm_ptr = get_ptr_to_shared_memory(shm_fd, sizeof(int));
+    void *shm_ptr = get_ptr_to_shared_memory(shm_fd, size);
 
     int err_code = *((int *) shm_ptr);
     assertm(err_code >= 0, "Error obtaining the shared memory pointer");
-    munmap(shm_ptr, sizeof(int));
+
+    shmem_close_shared_memory(TEST_SHM_NAME, shm_ptr, size);
+}
+
+void test_shmem_should_close() {
+    printf("### test_shmem_should_close\n");
+    int shm_fd = open_shared_memory(TEST_SHM_NAME);
+    size_t size = sizeof(int);
+    give_size_to_shmem(shm_fd, size);
+
+    void *shm_ptr = get_ptr_to_shared_memory(shm_fd, size);
+
+    int res = shmem_close_shared_memory(TEST_SHM_NAME, shm_ptr, size);
+    assertm(res==0, "Closing shared memory found an error");
+
+    int reopening_errcode = shm_open(TEST_SHM_NAME, O_RDWR, 0666);
+    assertm(reopening_errcode<0, "Shared memory could still be openned after supposedly closing it");
 }
 
 void test_should_write_various_structures_in_sh_mem() {
@@ -81,22 +88,19 @@ void test_should_write_various_structures_in_sh_mem() {
     memcpy(shm_ptr, &st1, sizeof(st1));
     memcpy(shm_ptr + sizeof(st1), &st2, sizeof(st2));
     memcpy(shm_ptr + sizeof(st1) + sizeof(st2), &st3, sizeof(st3));
-
-    munmap(shm_ptr, sizes);
 }
 
 void teardown() {
-    close_shared_memory(TEST_SHM_NAME);
     return;
 }
 
 int main() {
     printf("### Running writer tests!\n\n");
 
-    call_test_teardown(&test_shmget_should_return_fd, &teardown);
-    call_test_teardown(&test_shmem_should_close, &teardown);
-    call_test_teardown(&test_should_give_a_size_to_shmem, &teardown);
-    call_test_teardown(&test_should_retrieve_ptr_from_shmem, &teardown);
+    call_test(&test_shmget_should_return_fd);
+    call_test(&test_shmem_should_close);
+    call_test(&test_should_give_a_size_to_shmem);
+    call_test(&test_should_retrieve_ptr_from_shmem);
     call_test(&test_should_write_various_structures_in_sh_mem);
 
     printf("\n### All tests were succesfull!\n");
